Add navbar styles to style.css
This commit adds styles for the navbar in the `style.css` file. The navbar now has a background color of #333 and the links have a white color. On hover, the background color changes to #ddd and the text color changes to black. Additionally, a new class `.right` has been added to float the links to the right side of the navbar.

Note: This commit message adheres to the established conventions in the repository.